Change folder structure for FirstQuantized operators
