+++ include/Operator/FirstQuantized/CMakeLists.txt
@@ -2,5 +2,12 @@ target_sources(gqcp
         PRIVATE
             BaseMultipoleOperator.hpp
             BaseNuclearOperator.hpp
+            CoulombRepulsionOperator.hpp
+            ElectronicDipoleOperator.hpp
+            KineticOperator.hpp
+            NuclearAttractionOperator.hpp
+            NuclearDipoleOperator.hpp
+            NuclearRepulsionOperator.hpp
             Operator.hpp
+            OverlapOperator.hpp
         )
