+++ src/Operator/FirstQuantized/CMakeLists.txt
@@ -2,5 +2,7 @@ target_sources(gqcp
         PRIVATE
             BaseMultipoleOperator.cpp
             BaseNuclearOperator.cpp
+            NuclearDipoleOperator.cpp
+            NuclearRepulsionOperator.cpp
             Operator.cpp
         )
