+++ src/Operator/FirstQuantized/Operator.cpp
@@ -22,72 +22,7 @@ namespace GQCP {
 
 
 /*
- *  NUCLEAR REPULSION OPERATOR - PUBLIC METHODS
- */
-
-/**
- *  @return the scalar value of this nuclear repulsion operator
- */
-double NuclearRepulsionOperator::value() const {
-
-    const auto& nuclei = this->nuclear_framework.nucleiAsVector();
-
-    // Sum over every unique nucleus pair
-    double value {0.0};
-    const auto n_nuclei = this->nuclearFramework().numberOfNuclei();
-    for (size_t i = 0; i < n_nuclei; i++) {
-        for (size_t j = i + 1; j < n_nuclei; j++ ) {
-            const auto nucleus1 = nuclei[i];
-            const auto nucleus2 = nuclei[j];
-
-            // The internuclear repulsion energy (Coulomb) for every nucleus pair is Z1 * Z2 / |R1 - R2|
-            value += nucleus1.charge() * nucleus2.charge() / nucleus1.calculateDistance(nucleus2);
-        }
-    }
-
-    return value;
-}
-
-
-
-/*
- *  NUCLEAR DIPOLE OPERATOR - CONSTRUCTOR
- */
-
-/**
- *  @param nuclear_framework            the nuclear framework underlying a nuclear operator
- *  @param o                            the origin of the multipole
- */
-NuclearDipoleOperator::NuclearDipoleOperator(const NuclearFramework& nuclear_framework, const Vector<double, 3>& o) :
-    BaseNuclearOperator(nuclear_framework),
-    BaseMultipoleOperator(o)
-{}
-
-
-
-/*
- *  NUCLEAR DIPOLE OPERATOR - PUBLIC METHODS
- */
-
-/**
- *  @return the value of this nuclear dipole operator
- */
-Vector<double, 3> NuclearDipoleOperator::value() const {
-
-    Vector<double, 3> mu = Vector<double, 3>::Zero();
-
-    const auto& nuclei = this->nuclear_framework.nucleiAsVector();
-    for (const auto& nucleus : nuclei) {
-        mu += nucleus.charge() * nucleus.position();
-    }
-
-    return mu;
-}
-
-
-
-/*
- *  OPERATOR - PUBLIC STATIC METHODS
+ *  PUBLIC STATIC METHODS
  */
 
 /**
