+++ src/QCMethod/FCI.cpp
@@ -17,7 +17,7 @@
 //
 #include "QCMethod/FCI.hpp"
 
-#include "Operator/FirstQuantized/Operator.hpp"
+#include "Operator/FirstQuantized/NuclearRepulsionOperator.hpp"
 
 
 namespace GQCP {
